generators: wave indices in 'modelIntensity'
Replace the hard-coded vector of wave indices, which was just a
vector containing the numbers 0 to nmbWaves-2, with a list of wave
indices obtained from the fit result with a wave name pattern that
should match anything but 'flat'. This should not make a difference,
but is probably a bit more readable.